requirements.txt
flask==3.0.0
opencv-python==4.9.0.80
numpy>=1.24.2,<2.0.0
onnxruntime>=1.16.0
Pillow>=10.2.0
requests==2.31.0
scikit-learn
skl2onnx>=1.14.0
